+++ src/config/chain/types.ts
@@ -56,7 +56,7 @@ export interface ChainConfig {
   registryChainName: string;
   prettyName: string;
   logoUrl?: string;
-  networkType: "mainnet" | "testnet";
+  networkType: "mainnet" | "testnet" | "devnet" | "local";
   lcd: string;
   rpc: string;
   indexer: string;