+++ src/lib/pages/custom-network/edit/components/ExportNetworkConfig.tsx
@@ -1,0 +1,5 @@
+import { Flex } from "@chakra-ui/react";
+
+export const ExportNetworkConfig = () => {
+  return <Flex>Export Json</Flex>;
+};