+++ src/pages/[network]/custom-network/add/manual.tsx
@@ -1,0 +1,3 @@
+import { AddNetworkManual } from "lib/pages/custom-network/manual";
+
+export default AddNetworkManual;